+++ src/core/components/auth-item/auth-item.jsx
