+++ src/core/components/auth-item/index.js
@@ -1,0 +1,3 @@
+export { default as ApiKeyAuth } from "./auth-api-key"
+export { default as BasicAuth } from "./auth-basic"
+export { default as AuthItem } from "./auth-item"
