+++ src/core/components/auth-section/auth-section.jsx
