+++ src/core/components/auth-section/index.js
@@ -1,0 +1,5 @@
+export { default as Auths } from "./auth-section"
+export { default as AuthHeader } from "./auth-header"
+export { default as AuthRow } from "./auth-row"
+export { default as AuthFormRow } from "./auth-form-row"
+export { default as AuthError } from "./auth-error"
