+++ src/core/components/auth/index.js
@@ -1,0 +1,4 @@
+export { default as AuthorizeBtn } from "./auth-btn"
+export { default as AuthorizeOperationBtn } from "./auth-operation-btn"
+export { default as AuthorizationPopup } from "./auth-popup"
+
