+++ src/core/components/oauth2-item/index.js
@@ -1,0 +1,3 @@
+export { default as Oauth2 } from "./oauth2-item"
+export { default as Oauth2Form } from "./oauth2-form"
+export { default as Oauth2FormData } from "./oauth2-form-data"
