+++ src/core/components/oauth2-item/oauth2-form-data.jsx
